requirements.txt
black==25.1.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cryptography==45.0.6
github3.py==4.0.1
idna==3.10
isort==6.0.1
linkify-it-py==2.0.3
markdown-it-py==3.0.0
mdit-py-plugins==0.5.0
mdurl==0.1.2
msgspec==0.19.0
mypy==1.17.1
mypy_extensions==1.1.0
node-semver==0.9.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pur==7.3.3
pycparser==2.22
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.4
rich==14.1.0
six==1.17.0
termcolor==3.1.0
textual==5.3.0
trogon==0.6.0
typing_extensions==4.14.1
uc-micro-py==1.0.3
uritemplate==4.2.0
urllib3==2.5.0
